refactor(shared): 归还执行与断言工具所有权
