+++ src/eval-workflows/assertions/deterministic.ts
@@ -1,6 +1,6 @@
 import _Ajv from 'ajv';
 import type { ToolCallInfo } from '../../executors/contracts/trace.js';
-import type { SyncAssertionType } from '../assertion-types.js';
+import type { SyncAssertionType } from '../../shared/assertion-types.js';
 import type { Assertion } from '../../inputs/contracts/assertion.js';
 
 const Ajv = _Ajv.default ?? _Ajv;
