+++ src/eval-workflows/runtime-adapter/evaluators/assertion-common.ts
@@ -6,7 +6,7 @@ import {
   type SchemaIdentity,
 } from '../../../evaluation-core/contracts/index.js';
 import type { EvaluatorBindingValue } from '../../../evaluation-core/evaluation/index.js';
-import { resolveAssertionLayer } from '../../../shared/assertions/layers.js';
+import { resolveAssertionLayer } from '../../assertions/layers.js';
 import { assertionContractValidationError } from '../../../shared/sample-contract.js';
 import type { Assertion } from '../../../inputs/contracts/assertion.js';
