+++ src/eval-workflows/runtime-adapter/evaluators/execution-assertions.ts
@@ -22,7 +22,7 @@ import {
   deterministicAssertionInputSourceKinds,
   type DeterministicAssertionContext,
   type DeterministicAssertionInputSourceKind,
-} from '../../../shared/assertions/deterministic.js';
+} from '../../assertions/deterministic.js';
 import type { ToolCallInfo } from '../../../executors/contracts/trace.js';
 import type { Assertion } from '../../../inputs/contracts/assertion.js';
 import type { SameProcessEvaluatorImplementation } from '../adapters/same-process.js';
