+++ src/eval-workflows/runtime-adapter/evaluators/output-assertions.ts
@@ -17,7 +17,7 @@ import {
   assertionUsesOnlyOutput,
   createIsolatedDeterministicAssertionEvaluator,
   type DeterministicAssertionContext,
-} from '../../../shared/assertions/deterministic.js';
+} from '../../assertions/deterministic.js';
 import type { Assertion } from '../../../inputs/contracts/assertion.js';
 import type { SameProcessEvaluatorImplementation } from '../adapters/same-process.js';
 import {
