+++ src/observability/analysis/coverage-analyzer.ts
@@ -19,8 +19,8 @@ import {
 } from 'node:fs';
 import type { ToolCallInfo } from '../../executors/contracts/trace.js';
 import type { AnalysisEntry } from './contracts.js';
-import { toolCallQuery } from '../../shared/tool-search.js';
-import { isToolCallSuccess } from '../../shared/tool-call-status.js';
+import { toolCallQuery } from '../tool-search.js';
+import { isToolCallSuccess } from '../../executors/tool-call-status.js';
 
 export interface KnowledgeEntry {
   path: string;
