+++ test/eval-workflows/runtime-adapter/execution-assertions.test.ts
@@ -43,7 +43,7 @@ import {
   EXECUTION_AWARE_SYNC_ASSERTION_TYPE_NAMES,
   OUTPUT_ONLY_SYNC_ASSERTION_TYPE_NAMES,
   deterministicAssertionInputSourceKinds,
-} from '../../../src/shared/assertions/deterministic.js';
+} from '../../../src/eval-workflows/assertions/deterministic.js';
 import type { Assertion } from '../../../src/inputs/contracts/assertion.js';
 import {
   testRuntime,
