+++ test/eval-workflows/runtime-adapter/output-assertions.test.ts
@@ -39,7 +39,7 @@ import {
   EXECUTION_AWARE_SYNC_ASSERTION_TYPE_NAMES,
   OUTPUT_ONLY_SYNC_ASSERTION_TYPE_NAMES,
   createIsolatedDeterministicAssertionEvaluator,
-} from '../../../src/shared/assertions/deterministic.js';
+} from '../../../src/eval-workflows/assertions/deterministic.js';
 import type { Assertion } from '../../../src/inputs/contracts/assertion.js';
 import { testRuntime, validDefinition, validPolicy } from '../../evaluation-core/compiler/fixtures.js';
