+++ test/eval-workflows/runtime-adapter/rubric-judge.test.ts
@@ -47,7 +47,7 @@ import {
   getJudgePromptHash,
   JUDGE_SYSTEM_PROMPT,
 } from '../../../src/shared/llm-prompts/judge-prompts.js';
-import { buildJudgeTraceSummary } from '../../../src/shared/llm-prompts/judge-trace.js';
+import { buildJudgeTraceSummary } from '../../../src/grading/judge-trace.js';
 import type { ExecutorFn } from '../../../src/executors/contracts/ports.js';
 import { testRuntime, validDefinition, validPolicy } from '../../evaluation-core/compiler/fixtures.js';
