+++ test/executors/tool-call-status.test.ts
@@ -4,7 +4,7 @@ import {
   isToolCallFailure,
   isToolCallUnknown,
   toolCallStatus,
-} from '../../src/shared/tool-call-status.js';
+} from '../../src/executors/tool-call-status.js';
 
 describe('toolCallStatus', () => {
   it('keeps a missing legacy outcome unknown instead of fabricating a failure', () => {
